serverrouterrefactor: Change server to router
Prior to this change, the server type only had a method called
ListenAndServe which would actually stop us being able to use it with
TLS or other transport methods.

This change makes the server into just a router, which can then be
plugged into any other type of server.
