+++ router/apphandler.go
@@ -1,4 +1,4 @@
-package server
+package router
 
 import (
 	"encoding/json"
