+++ router/balancehandlers_test.go
@@ -1,4 +1,4 @@
-package server
+package router
 
 import (
 	"net/http"
@@ -14,7 +14,7 @@ import (
 func Test_balances(t *testing.T) {
 	t.Run("SelectAccount error", func(t *testing.T) {
 		expected := errors.New("account error")
-		srv := &server{
+		srv := &router{
 			storage: &storagetest.Storage{
 				AccountErr: expected,
 			},
@@ -28,7 +28,7 @@ func Test_balances(t *testing.T) {
 	t.Run("SelectBalance error", func(t *testing.T) {
 		account := &storage.Account{}
 		expected := errors.New("balances error")
-		srv := &server{
+		srv := &router{
 			storage: &storagetest.Storage{
 				Account:     account,
 				BalancesErr: expected,
@@ -42,7 +42,7 @@ func Test_balances(t *testing.T) {
 
 	t.Run("all ok", func(t *testing.T) {
 		expected := &storage.Balances{{ID: 1}}
-		srv := &server{
+		srv := &router{
 			storage: &storagetest.Storage{
 				Account:  &storage.Account{},
 				Balances: expected,
@@ -59,7 +59,7 @@ func Test_balances(t *testing.T) {
 func TestServer_InsertBalance(t *testing.T) {
 	t.Run("SelectAccount error", func(t *testing.T) {
 		expected := errors.New("SelectAccount error")
-		srv := server{&storagetest.Storage{
+		srv := router{&storagetest.Storage{
 			AccountErr: expected,
 		}}
 		code, b, err := srv.insertBalance(0, balance.Balance{})
@@ -71,7 +71,7 @@ func TestServer_InsertBalance(t *testing.T) {
 
 	t.Run("InsertBalance error", func(t *testing.T) {
 		expected := errors.New("InsertBalance error")
-		srv := server{&storagetest.Storage{
+		srv := router{&storagetest.Storage{
 			Account:    &storage.Account{},
 			BalanceErr: expected,
 		}}
@@ -84,7 +84,7 @@ func TestServer_InsertBalance(t *testing.T) {
 
 	t.Run("all ok", func(t *testing.T) {
 		expected := &storage.Balance{}
-		srv := server{&storagetest.Storage{
+		srv := router{&storagetest.Storage{
 			Account: &storage.Account{},
 			Balance: expected,
 		}}
