--- server/router.go
@@ -1,26 +1,0 @@
-package server
-
-import (
-	"log"
-
-	"github.com/gorilla/mux"
-	"github.com/pkg/errors"
-)
-
-// newRouter creates a new Router and initialises it will all of the global routes
-func (s *server) newRouter() (*mux.Router, error) {
-	if s == nil {
-		return nil, errors.New("nil server provided")
-	}
-	router := mux.NewRouter().StrictSlash(true)
-	for _, route := range s.routes() {
-		handler := logger(route.appHandler, route.name)
-		router.
-			Methods(route.method).
-			Path(route.pattern).
-			Name(route.name).
-			Handler(handler)
-		log.Printf("Route registered: %+v", route)
-	}
-	return router, nil
-}
